introduce POM

+++ tests/cleandb.ts
@@ -1,6 +1,6 @@
 import mariadb from 'mariadb';
 
-export async function cleandb() {
+export async function cleandb({ removeUsers = false } = {}) {
   const connection = await mariadb.createConnection({
     host: process.env.DB_HOST || 'localhost',
     user: process.env.DB_USERNAME || 'root',
@@ -13,7 +13,9 @@ export async function cleandb() {
   await connection.query(`DELETE FROM article;`);
   await connection.query(`DELETE FROM comment;`);
   await connection.query(`DELETE FROM tag;`);
-  await connection.query(`DELETE FROM user;`);
+  if (removeUsers) {
+    await connection.query(`DELETE FROM user;`);
+  }
   await connection.query(`DELETE FROM user_favorites;`);
   await connection.query(`DELETE FROM user_to_follower;`);
   await connection.commit();
